+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0 (Unreleased)
 
 - [improvement]: fractional string input can include integral part separated with whitespace: "1 2/3";
+- [feature]: new type of input, degrees/minutes/seconds string: (1.23'45'');
 
 ## 0.3.0 (2021-08-15)